refactor: Move Models to External/Database folder
