+++ src_temp/Infrastructure/External/Database/Models/Customer.ts
