+++ src_temp/Infrastructure/External/Database/Models/OrderItem.ts
