+++ src_temp/Infrastructure/External/Database/Models/Payment.ts
